UP-4706:  Moved NodeInfoTracker to it's own .java file;  a suggetion from the PR
